parser/pyproject.toml
[project]
name = "parser"
version = "0.1.0"
description = "Scientific paper acquisition, DOI to BibTeX conversion, and citation verification"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Jaime Cernuda" }
]
keywords = ["doi", "bibtex", "arxiv", "scientific-papers", "citation", "paper-download"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI & Output
    "click>=8.1.0",
    "rich>=13.0.0",
    # Config
    "pyyaml>=6.0",
    "pydantic>=2.0",
    # Async & HTTP
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "claude-agent-sdk>=0.1.18",
    "google-genai>=1.56.0",
    "google-adk>=1.21.0",
    "selenium>=4.39.0",
    "webdriver-manager>=4.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

# AI Agent Dependencies
# Claude Agent SDK (RECOMMENDED - no API key needed!)
claude-sdk = [
    "claude-agent-sdk>=0.1.0",
]

# Direct Anthropic API (requires ANTHROPIC_API_KEY)
anthropic = [
    "anthropic>=0.40.0",
]

# Google Gemini (requires GOOGLE_API_KEY)
gemini = [
    "google-generativeai>=0.8.0",
]

# Google ADK (optional, for advanced agent features)
google-adk = [
    "google-adk>=0.2.0",
]

# All AI agents
agents = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "google-adk>=0.2.0",
]
# Full installation with all optional dependencies
all = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "google-adk>=0.2.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
parser = "parser.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parser"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_return_any = false
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["yaml", "aiofiles"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "network: requires network access (run with --network)",
    "slow: marks tests as slow running",
]
